fixup! decouple ServicesKeyYamlToPhpFactory
