+++ phpstan.neon
@@ -40,3 +40,5 @@ parameters:
          - '#Variable \$methodCall might not be defined#'
          - '#Cognitive complexity for "Migrify\\ConfigTransformer\\FormatSwitcher\\PhpParser\\NodeFactory\\FluentClosureNamespaceNodeFactory\:\:createClosureStmts\(\)" is 9, keep it under 8#'
          - '#Parameter \#1 \$parameterName of method Migrify\\ConfigTransformer\\FormatSwitcher\\PhpParser\\NodeFactory\\PhpNodeFactory\:\:createParameterSetMethodCall\(\) expects string, int\|string given#'
+         - '#Cognitive complexity for "Migrify\\ConfigTransformer\\FormatSwitcher\\Converter\\KeyYamlToPhpFactory\\ServicesKeyYamlToPhpFactory\:\:convertServiceOptionsToNodes\(\)" is \d+, keep it under 8#'
+         - '#Cognitive complexity for "Migrify\\ConfigTransformer\\FormatSwitcher\\Converter\\KeyYamlToPhpFactory\\ServicesKeyYamlToPhpFactory\:\:convertYamlToNodes\(\)" is 9, keep it under 8#'
